Deduplicate code for running XPath|XQuery scripts
